Latest Revision of Gif Generation Javascript Code
The Code to query the Riffsy database works, however, even when set to synchronous, the next function to set the video url receives undefined for reasons unknown. Later the server eventually returns the json file as requested, but then the code has progressed past that, and the code fails.

I could not think of an easy way to rewrite the code to implement a waiting period for the complete URL to be returned for the videos. 


Otherwise, the gif objects are generated in either a HUNT mode or a PAIR mode.

Later, code can be added that when two gifs are selected, if their source url matches, then they are matches, so flip them over.

The Inner HTML to set the theme for HUNT is also broken for some reason.